pyproject.toml
[tool.poetry]
name = "mpa"
version = "0.1.0"
description = ""
authors = ["Dounx <imdounx@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
maafw = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
